+++ packages/core/styles/src/packages/composite/accordion/index.ts
@@ -1,0 +1,7 @@
+import * as styles from "./accordion.style.css";
+import { accordionSprinkle } from "./accordion.sprinkle.css";
+
+export const accordion = {
+  classnames: { ...styles },
+  ...accordionSprinkle,
+};