+++ packages/react/src/composite/Accordion/src/accordion.stories.tsx
@@ -1,0 +1,309 @@
+import React from "react";
+import { ComponentMeta, ComponentStory } from "@storybook/react";
+import { withA11y } from "@storybook/addon-a11y";
+import { TiendanubeIcon, UserGroupIcon } from "@tiendanube/icons";
+
+import { Text } from "@nimbus-ds/text";
+import { Icon } from "@nimbus-ds/icon";
+import { Box } from "@nimbus-ds/box";
+import { Stack } from "@nimbus-ds/stack";
+import { Card } from "@nimbus-ds/card";
+import { Title } from "@nimbus-ds/title";
+import { Radio } from "@nimbus-ds/radio";
+
+import { Accordion } from "./Accordion";
+
+export default {
+  title: "Composite/Accordion",
+  component: Accordion,
+  subcomponents: {
+    "Accordion.Header": Accordion.Header,
+    "Accordion.Body": Accordion.Body,
+    "Accordion.Item": Accordion.Item,
+  },
+
+  parameters: {
+    withA11y: { decorators: [withA11y] },
+  },
+  argTypes: {
+    children: { control: { disable: true } },
+  },
+} as ComponentMeta<typeof Accordion>;
+
+const Template: ComponentStory<typeof Accordion> = (args) => (
+  <Accordion {...args} />
+);
+
+export const base = Template.bind({});
+base.args = {
+  children: (
+    <>
+      <Accordion.Item index="0">
+        <Accordion.Header
+          title="Title one"
+          subtitle="Caption one"
+          icon={<TiendanubeIcon size={18} />}
+        />
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+      <Accordion.Item index="1">
+        <Accordion.Header
+          title="Title two"
+          subtitle="Caption two"
+          icon={<UserGroupIcon size={18} />}
+          borderBottom="base"
+        />
+        <Accordion.Body borderBottom="base">
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+    </>
+  ),
+};
+
+export const WithCard = Template.bind({});
+WithCard.args = {
+  children: (
+    <Card padding="none">
+      <Accordion.Item index="0">
+        <Accordion.Header
+          title="Title one"
+          subtitle="Caption one"
+          icon={<TiendanubeIcon size={18} />}
+          borderTop="none"
+        />
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+      <Accordion.Item index="1">
+        <Accordion.Header
+          title="Title two"
+          subtitle="Caption two"
+          icon={<UserGroupIcon size={18} />}
+        />
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+    </Card>
+  ),
+};
+
+export const withoutIcon = Template.bind({});
+withoutIcon.args = {
+  children: (
+    <Card padding="none">
+      <Accordion.Item index="0">
+        <Accordion.Header
+          title="Title one"
+          subtitle="Caption one"
+          borderTop="none"
+        />
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+      <Accordion.Item index="1">
+        <Accordion.Header title="Title two" subtitle="Caption two" />
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+    </Card>
+  ),
+};
+
+export const WithCustomIcon = Template.bind({});
+WithCustomIcon.args = {
+  children: (
+    <Card padding="none">
+      <Accordion.Item index="0">
+        <Accordion.Header
+          title="Title one"
+          subtitle="Caption one"
+          borderTop="none"
+        >
+          <Box
+            backgroundColor="primary.surface"
+            padding="2,5"
+            borderRadius="8px"
+            marginRight="2,5"
+          >
+            <Icon
+              color="neutral.textHigh"
+              source={<TiendanubeIcon size="medium" />}
+            />
+          </Box>
+        </Accordion.Header>
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+      <Accordion.Item index="1">
+        <Accordion.Header title="Title two" subtitle="Caption two">
+          <Box
+            backgroundColor="primary.surface"
+            padding="2,5"
+            borderRadius="8px"
+            marginRight="2,5"
+          >
+            <Icon
+              color="neutral.textHigh"
+              source={<UserGroupIcon size="medium" />}
+            />
+          </Box>
+        </Accordion.Header>
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+    </Card>
+  ),
+};
+
+export const withoutIconToogle = Template.bind({});
+withoutIconToogle.args = {
+  children: (
+    <Card padding="none">
+      <Accordion.Item index="0">
+        <Accordion.Header noIconToggle borderTop="none">
+          {({ selected }) => (
+            <Stack
+              width="100%"
+              display="flex"
+              justifyContent="space-between"
+              flexShrink="0"
+            >
+              <Title as="h5">Lorem Ipsum one</Title>
+              <Radio name="accordion-radio-0" checked={selected === "0"} />
+            </Stack>
+          )}
+        </Accordion.Header>
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+      <Accordion.Item index="1">
+        <Accordion.Header noIconToggle>
+          {({ selected }) => (
+            <Stack
+              width="100%"
+              display="flex"
+              justifyContent="space-between"
+              flexShrink="0"
+            >
+              <Title as="h5">Lorem Ipsum two</Title>
+              <Radio name="accordion-radio-0" checked={selected === "1"} />
+            </Stack>
+          )}
+        </Accordion.Header>
+        <Accordion.Body>
+          <Box
+            width="100%"
+            borderStyle="dashed"
+            padding="2"
+            borderWidth="1px"
+            borderColor="neutral.interactive"
+          >
+            <Text textAlign="center" fontSize="base">
+              Replace me with your content
+            </Text>
+          </Box>
+        </Accordion.Body>
+      </Accordion.Item>
+    </Card>
+  ),
+};