+++ packages/react/src/composite/Accordion/src/components/AccordionBody/index.ts
@@ -1,0 +1,4 @@
+import { AccordionBody } from "./AccordionBody";
+
+export { AccordionBody } from "./AccordionBody";
+export default AccordionBody;