+++ packages/react/src/composite/Accordion/src/components/AccordionHeader/index.ts
@@ -1,0 +1,4 @@
+import { AccordionHeader } from "./AccordionHeader";
+
+export { AccordionHeader } from "./AccordionHeader";
+export default AccordionHeader;