+++ packages/react/src/composite/Accordion/src/contexts/AccordionContext/AccordionContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { AccordionContextProps } from "./accordionContext.types";
+
+export const AccordionContext = createContext<AccordionContextProps>(
+  {} as AccordionContextProps
+);