+++ packages/react/src/composite/Accordion/src/contexts/AccordionContext/index.ts
@@ -1,0 +1,5 @@
+import { AccordionContext } from "./AccordionContext";
+
+export { AccordionContext } from "./AccordionContext";
+export type { AccordionContextProps } from "./accordionContext.types";
+export default AccordionContext;