+++ packages/react/src/composite/Accordion/src/contexts/AccordionItemContext/index.ts
@@ -1,0 +1,5 @@
+import { AccordionItemContext } from "./AccordionItemContext";
+
+export { AccordionItemContext } from "./AccordionItemContext";
+export type { AccordionItemContextProps } from "./accordionItemContext.types";
+export default AccordionItemContext;