+++ packages/react/src/composite/Accordion/src/hooks/useAccordionItem/useAccordionItem.ts
@@ -1,0 +1,17 @@
+import { useContext } from "react";
+
+import {
+  AccordionItemContext,
+  AccordionItemContextProps,
+} from "../../contexts";
+
+export const useAccordionItem: () => AccordionItemContextProps = () => {
+  const context = useContext(AccordionItemContext);
+
+  if (context == null || Object.keys(context).length === 0) {
+    throw new Error(
+      "AccordionHeader or AccordionBody components must be wrapped in <AccordionItem />"
+    );
+  }
+  return context;
+};